pyproject.toml
[project]
name = "xai-final-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "joblib>=1.5.3",
    "keras>=3.13.0",
    "librosa>=0.11.0",
    "lime>=0.2.0.1",
    "matplotlib>=3.10.8",
    "numpy>=2.3.5",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "scikit-learn>=1.8.0",
    "scipy>=1.16.3",
    "seaborn>=0.13.2",
    "shap>=0.49.1",
    "tensorflow>=2.20.0",
    "tensorflow-keras>=0.1",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "python-multipart>=0.0.21",
]
